Start on fetching historic data from Lichess
Lines out simple controller/model/service/utility folder
structure. Not set in stone. Create a thin controller that
is responsible for returning Lichess entities, and a service
that does the lifting.

LichessApi wraps up the communication with Lichess, and
HttpUtility encapsulates the detail of making HTTP requests.
The idea here is that we currently want some kind of throttling
so we don't get an IP ban, but that logic may change in the future.