Adds request caching at the HTTP request level
Requests made via the ThrottledHttpUtility are now
cached. Every 30 minutes a scheduled task runs to
clear the cache.

Rather than using a cache provider, I decided to keep
it simple and let Spring back the cache with a
ConcurrentHashMap. This means we don't get things like
timed cache expiry, but it seemed a bit overkill to
use something like Infinispan at this point.

Note the lack of a static final for the name of the cache
in ThrottledHttpUtility & see
(GROOVY-3278)[https://issues.apache.org/jira/browse/GROOVY-3278].
As discussed the Groovy compiler expects an inline constant, but
it looks like Groovy uses a PropertyExpression to represent class
fields (or something along those lines).