Spring Cache now uses Google Guava cache implementation
Also note the removal of the chessmates.utiltiy.Cache interface and
the removal of the evict method from the ThrottledHttpUtility. We now
clear the cache in the CacheController by fetching the cache from the
`CacheManager` by name.

Also, for some unknown reason, the Groovy compiler doesn't moan when
class fields are used in annotations when they are defined on another
class (compilation unit?). Since I've moved the actual cache creation into
Application, I now define the cache name as a static final String there too.
Meaning I can use that property. So strange.