Use JSON files for unit tests instead of generating in code
This adds a bunch of mock Lichess responses, 2 pages for 3 users
and two pages of team member data. The page sizes in the tests are
set intentionally small so that we get paged results for the small team we
have.

The page sizes have been pulled out into application.properties, this allows
us to inject them in our test env, but is nicer anyway as a bonus. The page sizes
are also now a responsibility of the LichessDataService, and not the LichessApi
which now takes page size as a parameter.