Removes sleep and throws
This currently just logs. At the moment continueing after the sleep
doesn't work. As the thread wakes up, it tries to open a reader on the
input stream which throws an IOException. Let's just log an
error and rethrow the error. We'll continue where we picked
up from next time we schedule the task. That'll do for now.