Fixes unit tests for LichessDataService
Since we weren't parsing to our model objects anymore
we want another way of validating the data (we just check that
id isn't null).

The lichess API just returns whatever and it's no up to the data
service to validate.

Tests needed to be updated to use closures for matching rather
than relying on our model objects .equals() implementation.