Adds getPlayer & getGame endpoint
I haven't added any tests here as theres no real logic. I've tested
the fetching of teams from the LichessDataService - that's the
real meat of the application so far. Is it worth adding in e2e tests
that mock out the repositories and hit the REST end points? Does
that add any real value? Opinions please!