Renames data service methods
Makes it a little clearer that these methods have
the effect of actually saving the data. I don't
like the way the LichessDataServie is responsible for:
- Getting the items
- Persisting the items
- Looking up the latest items
- Returning the list of items from each method

It makes it very unclear that there are side effects of
persisting the data when looking at the Interface. I think
the LichessDataService should be refactored to simply
return the objects, and the LichessDataScheduler should
look at persisting the items & looking up the latest items
fetched. I'll raise this as a GH issue.